Rename daqspectrum file to spectrum
We can now take spectra with Zurich as well!
