+++ Measurements/scanspectra.py
@@ -11,7 +11,7 @@ from ..Utilities.plotting import plot_mpl
 from ..Instruments import piezos, montana, squidarray
 from .measurement import Measurement
 from ..Utilities import conversions
-from ..Measurements.daqspectrum import SQUIDSpectrum
+from ..Measurements.spectrum import SQUIDSpectrum
 from ..Utilities.utilities import AttrDict
 
 class Scanspectra(Measurement):
