+++ Measurements/spectrum.py
